backend-cal/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
gunicorn>=20.1.0
djangorestframework
djangorestframework-simplejwt


# dev_dependencies
django-extensions
pydotplus

